refactor(inspector): extract EffectsTab + ParticleEffectsSectionWrapper
